+++ led_drawables.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <Arduino.h>
+#include <FastLED.h>
 class Canvas;
 
 //TODO: Figure out how to delete class on command
@@ -7,42 +9,42 @@ class Drawable
   public:
     //Accessors:
     //Position:
-    int8_t get_pos_x();
-    int8_t get_pos_y();
-    int8_t get_pos(); //Gets [pos_x,pos_y]
+    int get_pos_x();
+    int get_pos_y();
+    int get_pos(); //Gets [pos_x,pos_y]
     //Size
-    uint8_t get_width();
-    uint8_t get_height();
-    uint8_t get_size(); //Gets [width,height]
+    int get_width();
+    int get_height();
+    int get_size(); //Gets [width,height]
     //Misc:
     int get_alpha();
     unsigned long get_age();
-    uint8_t inbounds(int x, int y);
+    int inbounds(int x, int y);
   
     //Modifiers
-    void set_pos_x(int8_t x);
-    void set_pos_y(int8_t y);
-    void set_visible(uint8_t);
-    void moveTo(int8_t x, int8_t y);
+    void set_pos_x(int x);
+    void set_pos_y(int y);
+    void set_visible(int);
+    void moveTo(int x, int y);
     void draw(Canvas *t_canvas); //Draws the shape on the canvas - different behavior for different shapes
     void ageUp();  //Increments age by 1
     void ageReset(); //Sets age back to 0
 
     protected:
-      int8_t pos_x;
-      int8_t pos_y;
-      uint8_t width;
-      uint8_t height;
+      int pos_x;
+      int pos_y;
+      int width;
+      int height;
       unsigned long age = 0;
-      uint8_t alpha;
-      uint8_t visible = 1;
-      uint8_t priority = 0;
+      int alpha;
+      int visible = 1;
+      int priority = 0;
 };
 
 class Point: public Drawable
 {
   public:
-    Point(int8_t x, int8_t y);
+    Point(int x, int y);
   
   private:
 };
@@ -51,8 +53,8 @@ class Point: public Drawable
 class Rect: public Drawable
 {
   public:
-    Rect(int8_t x, int8_t y, uint8_t s_x, uint8_t s_y);
-    Rect(int8_t x, int8_t y, uint8_t s_x, uint8_t s_y, uint8_t t_alpha);
+    Rect(int x, int y, int s_x, int s_y);
+    Rect(int x, int y, int s_x, int s_y, int t_alpha);
     void draw(Canvas *t_c);
   
   private:
@@ -61,12 +63,12 @@ class Rect: public Drawable
 class Line: public Drawable
 { 
   public:
-    Line(int8_t x, int8_t y, int8_t x2, int8_t y2);
-    Line(int8_t x, int8_t y, int8_t x2, int8_t y2, uint8_t t_alpha);
+    Line(int x, int y, int x2, int y2);
+    Line(int x, int y, int x2, int y2, int t_alpha);
 
   private:
-    int8_t pos_x2;
-    int8_t pos_y2;
+    int pos_x2;
+    int pos_y2;
 };
 
 class DrawableChar: public Drawable
@@ -89,9 +91,9 @@ class DrawableString: public Drawable
 class Sprite: public Drawable
 {
   public:
-    char changeSprite(uint8_t *n_Sprite);
+    char changeSprite(int *n_Sprite);
 
   private:
-    uint8_t* sprite;
+    int* sprite;
     CHSV palette[];
 };
